+++ Modules/Intent.Modules.Entities/OperationExtensions.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Intent.Metadata.Models;
-using Intent.Modelers.Domain.Api;
+﻿using Intent.Modelers.Domain.Api;
 using Intent.Modules.Common;
 
 namespace Intent.Modules.Entities
@@ -13,7 +7,7 @@ namespace Intent.Modules.Entities
     {
         public static bool IsAsync(this OperationModel operation)
         {
-            return operation.HasStereotype("Asynchronous");
+            return operation.HasStereotype("Asynchronous") || operation.Name.EndsWith("Async");
         }
     }
 }